requirements.txt
# Core LangChain (latest compatible set)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# FIX: Use latest langchain-groq
langchain-groq>=0.3.8

# FIX: Downgrade groq (IMPORTANT)
groq>=0.30.0,<1.0.0

# Tavily
tavily-python>=0.3.0

# Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Env
python-dotenv>=1.0.0

# Async
aiohttp>=3.9.0

# Data
pandas>=2.0.0

# Token utils
tiktoken>=0.6.0

# Logging
rich>=13.7.0

# Retry
tenacity>=8.2.0

# JSON
orjson>=3.9.0

# Typing
pydantic>=2.5.0

# HTML parsing
html5lib>=1.1

streamlit>=1.30.0